Re-organize server code, improve benchmark
