+++ server/src/handlers/mod.rs
@@ -1,5 +1,4 @@
 pub mod commit;
-pub mod home;
 pub mod resource;
 pub mod single_page_app;
 pub mod tpf;
