+++ server/src/main.rs
@@ -1,5 +1,5 @@
 use actix_cors::Cors;
-use actix_web::{App, HttpServer, http::{Method}, middleware, web};
+use actix_web::{middleware, web, App, HttpServer};
 use std::{io, sync::Mutex};
 mod appstate;
 mod config;
@@ -10,6 +10,9 @@ mod helpers;
 mod https;
 mod jsonerrors;
 mod render;
+mod routes;
+#[cfg(feature = "desktop")]
+mod tray_icon;
 mod views;
 
 #[actix_rt::main]
@@ -31,89 +34,26 @@ async fn main() -> io::Result<()> {
         }
     };
 
-    let server =
-        HttpServer::new(move || {
-            let data = web::Data::new(Mutex::new(appstate.clone()));
-            // Allow requests from other domains
-            let cors = Cors::permissive();
+    let server = HttpServer::new(move || {
+        let data = web::Data::new(Mutex::new(appstate.clone()));
+        // Allow requests from other domains
+        let cors = Cors::permissive();
 
-            App::new()
-                .app_data(data)
-                .wrap(cors)
-                .wrap(middleware::Logger::default())
-                .wrap(middleware::Compress::default())
-                // Catch all HTML requests and send them to the single page app
-                .service(
-                web::resource("/*")
-                    .guard(actix_web::guard::Method(Method::GET))
-                    .guard(actix_web::guard::fn_guard(|head|
-                        content_types::get_accept(&head.headers()) == content_types::ContentType::HTML ))
-                    .to(handlers::single_page_app::single_page)
-                )
-                .service(actix_files::Files::new("/static", "static/").show_files_listing())
-                .service(
-                    actix_files::Files::new("/.well-known", "static/well-known/")
-                        .show_files_listing(),
-                )
-                .service(
-                    web::scope("/tpf")
-                        .service(web::resource("").route(web::get().to(handlers::tpf::tpf))),
-                )
-                .service(web::scope("/commit").service(
-                    web::resource("").route(web::post().to(handlers::commit::post_commit)),
-                ))
-                .service(
-                    web::scope("/validate").service(
-                        web::resource("").route(web::get().to(handlers::validate::validate)),
-                    ),
-                )
-                .service(web::scope("/{path:[^{}]+}").service(
-                    web::resource("").route(web::get().to(handlers::resource::get_resource)),
-                ))
-                .service(
-                    web::scope("/")
-                        .service(web::resource("").route(web::get().to(handlers::home::home))),
-                )
-                .app_data(
-                    web::JsonConfig::default()
-                        // register error_handler for JSON extractors.
-                        .error_handler(jsonerrors::json_error_handler),
-                )
-        });
+        App::new()
+            .app_data(data)
+            .wrap(cors)
+            .wrap(middleware::Logger::default())
+            .wrap(middleware::Compress::default())
+            .configure(routes::config_routes)
+            .app_data(
+                web::JsonConfig::default()
+                    // register error_handler for JSON extractors.
+                    .error_handler(jsonerrors::json_error_handler),
+            )
+    });
 
-    // Add tray icon
-    // If this turns out to suck, remove the tray_icon and webbrowser crates
-    #[cfg(feature="desktop")]
-    {
-        let server_url = if config.https {
-            format!("https://{}:{}", config.ip.to_string(), config.port_https)
-        } else {
-            format!("http://{}:{}", config.ip.to_string(), config.port)
-        };
-        actix_rt::spawn(async move {
-            let mut tray = match tray_item::TrayItem::new("Atomic", "") {
-                Ok(item) => item,
-                Err(_e) => return,
-            };
-            let _ = tray.add_menu_item("Open in browser", move || {
-                if webbrowser::open(&server_url.clone()).is_ok() {
-                    log::info!("Opening browser url...")
-                } else {
-                    log::info!("Opening browser url failed.")
-                }
-            });
-            let _ = tray.add_menu_item("About", move || {
-                if webbrowser::open("https://github.com/joepio/atomic").is_ok() {
-                    log::info!("Opening about url...")
-                } else {
-                    log::info!("Opening about url failed.")
-                }
-            });
-            let inner = tray.inner_mut();
-            inner.add_quit_item("Quit");
-            inner.display();
-        });
-    }
+    #[cfg(feature = "desktop")]
+    tray_icon::tray_icon_process();
 
     if config.https {
         // If there is no certificate file, or the certs are too old, start HTTPS initialization
